Componentizado graficos
