+++ src/app/graficos/grafico-area/grafico-area.component.html
@@ -1,0 +1,10 @@
+<div class="flex-box">
+    <apx-chart 
+      [series]="chartOptions2.series" 
+      [chart]="chartOptions2.chart" 
+      [xaxis]="chartOptions2.xaxis"
+      [stroke]="chartOptions2.stroke" 
+      [tooltip]="chartOptions2.tooltip" 
+      [dataLabels]="chartOptions2.dataLabels"
+    ></apx-chart>
+</div>
