+++ src/app/graficos/grafico-barras/grafico-barras.component.html
@@ -1,0 +1,8 @@
+<div class="flex-box">
+    <apx-chart 
+        [series]="chartOptions.series" 
+        [chart]="chartOptions.chart" 
+        [xaxis]="chartOptions.xaxis"
+        [title]="chartOptions.title"
+    ></apx-chart>
+</div>
