+++ src/app/graficos/grafico-barras/grafico-barras.component.ts
@@ -1,0 +1,61 @@
+import { Component, OnInit, ViewChild } from '@angular/core';
+import {
+  ChartComponent,
+  ApexAxisChartSeries,
+  ApexChart,
+  ApexXAxis,
+  ApexTitleSubtitle,
+} from 'ng-apexcharts';
+
+export type ChartOptions = {
+  series: ApexAxisChartSeries;
+  chart: ApexChart;
+  xaxis: ApexXAxis;
+  title: ApexTitleSubtitle;
+};
+
+@Component({
+  selector: 'app-grafico-barras',
+  templateUrl: './grafico-barras.component.html',
+  styleUrls: ['./grafico-barras.component.scss']
+})
+export class GraficoBarrasComponent implements OnInit {
+  @ViewChild('chart')
+  chart: ChartComponent = new ChartComponent();
+  public chartOptions: ChartOptions;
+
+  constructor() {
+    this.chartOptions = {
+      series: [
+        {
+          name: 'My-series',
+          data: [10, 41, 35, 51, 49, 62, 69, 91, 148],
+        },
+      ],
+      chart: {
+        height: 350,
+        type: 'bar',
+      },
+      title: {
+        text: 'Angular Chart',
+      },
+      xaxis: {
+        categories: [
+          'Jan',
+          'Feb',
+          'Mar',
+          'Apr',
+          'May',
+          'Jun',
+          'Jul',
+          'Aug',
+          'Sep',
+        ],
+      },
+    };
+   }
+
+  ngOnInit(): void {
+  }
+
+}
