+++ src/app/graficos/grafico-candle/grafico-candle.component.html
@@ -1,0 +1,9 @@
+<div class="flex-box">
+    <apx-chart
+        [series]="chartOptions4.series"
+        [chart]="chartOptions4.chart"
+        [xaxis]="chartOptions4.xaxis"
+        [yaxis]="chartOptions4.yaxis"
+        [title]="chartOptions4.title"
+    ></apx-chart>
+</div>
