+++ src/app/graficos/grafico-pizza/grafico-pizza.component.html
@@ -1,0 +1,8 @@
+<div class="flex-box">
+    <apx-chart
+        [series]="chartOptions3.series"
+        [chart]="chartOptions3.chart"
+        [labels]="chartOptions3.labels"
+        [responsive]="chartOptions3.responsive"
+    ></apx-chart>
+</div>
