+++ src/app/graficos/grafico-pizza/grafico-pizza.component.ts
@@ -1,0 +1,53 @@
+import { Component, OnInit, ViewChild } from '@angular/core';
+import {
+  ChartComponent,
+  ApexNonAxisChartSeries,
+  ApexResponsive,
+  ApexChart,
+} from 'ng-apexcharts';
+
+export type ChartOptions3 = {
+  series: ApexNonAxisChartSeries;
+  chart: ApexChart;
+  responsive: ApexResponsive[];
+  labels: any;
+};
+
+@Component({
+  selector: 'app-grafico-pizza',
+  templateUrl: './grafico-pizza.component.html',
+  styleUrls: ['./grafico-pizza.component.scss']
+})
+export class GraficoPizzaComponent implements OnInit {
+  @ViewChild("chart3")
+  chart3: ChartComponent = new ChartComponent;
+  public chartOptions3: ChartOptions3;
+
+  constructor() {
+    this.chartOptions3 = {
+      series: [44, 55, 13, 43, 22],
+      chart: {
+        width: 380,
+        type: "pie"
+      },
+      labels: ["Team A", "Team B", "Team C", "Team D", "Team E"],
+      responsive: [
+        {
+          breakpoint: 480,
+          options: {
+            chart: {
+              width: 200
+            },
+            legend: {
+              position: "bottom"
+            }
+          }
+        }
+      ]
+    };
+   }
+
+  ngOnInit(): void {
+  }
+
+}
